server/requirements.txt
alembic==1.0.8
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
more-itertools==7.0.0
pip-autoremove==0.9.1
pluggy==0.9.0
py==1.8.0
pytest==4.4.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
WTForms==2.2.1
